[ MOSIP-25231 ] updated install.sh & delete.sh for partner-onboarder

+++ deployment/v3/mosip/partner-onboarder/delete.sh
@@ -6,6 +6,12 @@ if [ $# -ge 1 ] ; then
   export KUBECONFIG=$1
 fi
 NS=onboarder
-echo Deleting partner-onboarder helm
-kubectl -n $NS delete configmap global
-helm -n $NS delete partner-onboarder
+while true; do
+    read -p "Are you sure you want to delete all partner-onboarder ?(Y/n) " yn
+    if [ $yn = "Y" ]; then
+      echo Deleting partner-onboarder helm
+      kubectl -n $NS --ignore-not-found=true  delete configmap global
+      helm -n $NS delete partner-onboarder
+      break
+    fi
+done
